requirements.txt
# Streamlit
streamlit>=1.28.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.0

# Embeddings and Vector Store
chromadb>=0.4.18
sentence-transformers>=2.2.2

# BM25 Search
rank-bm25>=0.2.2

# LLM Integration
openai>=1.3.0
anthropic>=0.7.0

# Text Processing
tiktoken>=0.5.1
langchain-text-splitters>=0.0.1

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
